+++ jgclark.Reviews/src/projectClass.js
@@ -2,7 +2,7 @@
 //-----------------------------------------------------------------------------
 // Project class definition for Review plugin
 // by Jonathan Clark
-// Last updated 2026-02-14 for v1.3.0.b10, @jgclark
+// Last updated 2026-02-16 for v1.3.0.b12, @jgclark
 //-----------------------------------------------------------------------------
 
 // Import Helper functions
@@ -34,7 +34,7 @@ import {
   getInputTrimmed,
   inputIntegerBounded,
 } from '@helpers/userInput'
-import { isClosedTask, isClosed, isOpen, isOpenTask } from '@helpers/utils'
+import { isClosedTask, isClosed, isOpen, isOpenTask, isOpenChecklist } from '@helpers/utils'
 
 //-----------------------------------------------------------------------------
 // Types
@@ -97,11 +97,11 @@ export class Project {
   completedDuration: ?string // string description of time to completion, or how long ago completed
   cancelledDate: ?Date
   cancelledDuration: ?string // string description of time to cancellation, or how long ago cancelled
-  numOpenItems: number
-  numCompletedItems: number
-  numTotalItems: number
-  numWaitingItems: number
-  numFutureItems: number
+  numOpenItems: number = 0
+  numCompletedItems: number = 0
+  numTotalItems: number = 0
+  numWaitingItems: number = 0
+  numFutureItems: number = 0
   isCompleted: boolean = false
   isCancelled: boolean = false
   isPaused: boolean = false